% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_helpers.R
\name{and_join}
\alias{and_join}
\title{Join a vector with 'AND'}
\usage{
and_join(s, leading_and = FALSE, trailing_and = FALSE)
}
\arguments{
\item{s}{a character vector}

\item{leading_and}{logical indicating whether 'and' should be pre-pended}

\item{trailing_and}{logical indicating whether 'and' should be post-pended}
}
\value{
a string with the elements of s joined by 'AND'
}
\description{
Join a vector with 'AND'
}
\examples{
and_join("")
and_join("country = 'US'")
and_join(c("country = 'US'", "type = 1"))
}
