% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_helpers.R
\name{blank_if_null}
\alias{blank_if_null}
\title{Insert blank character if a variable is null}
\usage{
blank_if_null(s)
}
\arguments{
\item{s}{String to insert, or blank if null

This is useful for specifying e.g. a column list that may or may not
be part of a query.}
}
\value{
a character vector
}
\description{
Insert blank character if a variable is null
}
