% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_helpers.R
\name{comma_join_if_not_null}
\alias{comma_join_if_not_null}
\title{Join a character vector}
\usage{
comma_join_if_not_null(
  s,
  leading_comma = TRUE,
  trailing_comma = FALSE,
  quote = FALSE
)
}
\arguments{
\item{s}{vector of items to join}

\item{leading_comma}{should a leading comma be inserted?}

\item{trailing_comma}{should a trailing comma be inserted?}

\item{quote}{should elements of the output be quoted?

This is useful for specifying e.g. a column list that may or may not
be part of a query.}
}
\value{
a character vector
}
\description{
Join a character vector
}
\examples{
comma_join_if_not_null(NULL) # returns ""
comma_join_if_not_null(1:3, leading_comma = FALSE) # returns "1, 2, 3"
}
