% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{env_from_params}
\alias{env_from_params}
\title{Creates the environment for using glue to replace}
\usage{
env_from_params(params, ...)
}
\arguments{
\item{params}{list of parameters}

\item{...}{filled parameter values}
}
\value{
an environment for gluing
}
\description{
Creates the environment for using glue to replace
}
\keyword{internal}
