% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{head.query_template}
\alias{head.query_template}
\title{Print the header of a query}
\usage{
\method{head}{query_template}(q)
}
\arguments{
\item{q}{a query}
}
\value{
silently returns the head
}
\description{
Print the header of a query
}
