% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{list_queries}
\alias{list_queries}
\title{List the available queries}
\usage{
list_queries(location = queries_location())
}
\arguments{
\item{location}{the directory to look in}
}
\value{
A tibble of queries
}
\description{
List the available queries
}
