% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_query}
\alias{load_query}
\title{Load a query from file}
\usage{
load_query(query_name, location = queries_location())
}
\arguments{
\item{query_name}{name of the query}

\item{location}{location of the query}
}
\value{
loaded query
}
\description{
Load a query from file
}
