% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_helpers.R
\name{named_vec_to_cols}
\alias{named_vec_to_cols}
\title{Use a named vector to define columns}
\usage{
named_vec_to_cols(s)
}
\arguments{
\item{s}{a named vector}
}
\value{
a character vector
}
\description{
Use a named vector to define columns
}
\examples{
named_vec_to_cols(c(average_sales = "avg(sales)", user_id = "userid"))
}
