% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{new_query}
\alias{new_query}
\title{Create a query file}
\usage{
new_query(
  name,
  description = NULL,
  params = NULL,
  body = NULL,
  location = queries_location()
)
}
\arguments{
\item{name}{name of the query}

\item{description}{description}

\item{params}{parameters for the query}

\item{body}{the select statement}

\item{location}{where to save?}
}
\description{
Create a query file
}
