% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{open_query}
\alias{open_query}
\title{Opens a query file in the editor}
\usage{
open_query(query_name, location = queries_location())
}
\arguments{
\item{query_name}{the name of the query}

\item{location}{the query directory}
}
\description{
Opens a query file in the editor
}
