% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{print.query_template}
\alias{print.query_template}
\title{Print query}
\usage{
\method{print}{query_template}(x, ...)
}
\arguments{
\item{x}{a query of time query_template}

\item{...}{ignored}
}
\description{
Print query
}
