% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qt.R
\name{qt_basic}
\alias{qt_basic}
\title{The most basic query pattern}
\usage{
qt_basic(
  select_cols,
  table_name,
  where = NULL,
  group_by = NULL,
  order_by = NULL,
  limit = NULL
)
}
\arguments{
\item{select_cols}{vector of columns to select. 
Names will be used as identifiers, if supplied}

\item{table_name}{table to select from}

\item{where}{vector of constraints to be joined with 'and'}

\item{group_by}{vector of columns to group by}

\item{order_by}{vector of columns to order by}

\item{limit}{number of rows to return}
}
\value{
a filled query
}
\description{
The most basic query pattern
}
\examples{
qt_basic(
   select_cols = c(Date = "saledate", "Country", Sales = "SUM(sales)"), 
   table = "all_sales",
   where = c("country in ('US', 'CA')", "year = 2020"),
   group_by = c("saledate", "country"),
   order_by = "Date")
}
