% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qt.R
\name{qt_rollup}
\alias{qt_rollup}
\title{Dimensional rollup query pattern}
\usage{
qt_rollup(dimensions, metrics, table_name, where = NULL, order_by = dimensions)
}
\arguments{
\item{dimensions}{dimensions to roll up by}

\item{metrics}{aggregations that will be performed on each dimension}

\item{table_name}{table to query}

\item{where}{constraints}

\item{order_by}{order by}
}
\value{
a query
}
\description{
Produces a query that performs a rollup operation on a cube-like table.
}
\examples{
qt_rollup(
  c(order_date = "date", "country"), 
  c(sales = "sum(sales)"), 
  'metrics_table'
  )
}
