% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{queries_default_location}
\alias{queries_default_location}
\title{Returns the default query location}
\usage{
queries_default_location()
}
\value{
default query location as a character
}
\description{
First checks for a queries_location option.
If that's not there, defualts to ~/.queries
}
