% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{queries_list}
\alias{queries_list}
\title{List the available queries}
\usage{
queries_list(location = queries_default_location())
}
\arguments{
\item{location}{the directory to look in}
}
\value{
A tibble of queries
}
\description{
List the available queries
}
