% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_as_function}
\alias{query_as_function}
\title{Turn a parameterized query into a function}
\usage{
query_as_function(x, include_header = FALSE, append_params = FALSE)
}
\arguments{
\item{x}{a query loaded with query_load}

\item{include_header}{include the header in the query output}

\item{append_params}{append parameter values in the query}
}
\value{
a function whose arguments are the query parameters
}
\description{
Turn a parameterized query into a function
}
