% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_create}
\alias{query_create}
\title{Create a parameterized query}
\usage{
query_create(filename, query_name = "", param_names = NULL, open = TRUE)
}
\arguments{
\item{filename}{what do you want the query to be called}

\item{query_name}{the name at the top of the query description}

\item{param_names}{names of query parameters}

\item{open}{if true, the saved file opens in rstudio}
}
\value{
invisibly
}
\description{
Create a parameterized query
}
