% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_from_string}
\alias{query_from_string}
\title{Build a query object from a length-1 character vector}
\usage{
query_from_string(s, include_header = FALSE)
}
\arguments{
\item{s}{character containing query}

\item{include_header}{if true, the header will be included in the sql query}
}
\value{
convert string into query object
}
\description{
Build a query object from a length-1 character vector
}
