% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_load}
\alias{query_load}
\title{Load a query from file}
\usage{
query_load(
  query_name,
  query_location = queries_default_location(),
  include_header = FALSE
)
}
\arguments{
\item{query_name}{name of the query. This can also be a path, in which
case query_location is ignored.}

\item{query_location}{location of the query.}

\item{include_header}{if true, the header will be included in the sql query}
}
\value{
loaded query
}
\description{
Load a query from file
}
