% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_set_default_location}
\alias{query_set_default_location}
\title{Sets the default queries location}
\usage{
query_set_default_location(path)
}
\arguments{
\item{path}{file path to a directory with .sql files}
}
\description{
Sets the default queries location
}
