% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_substitute}
\alias{query_substitute}
\title{Replace parameters in query}
\usage{
query_substitute(
  qry,
  ...,
  query_location = queries_default_location(),
  include_header = FALSE,
  append_params = TRUE
)
}
\arguments{
\item{qry}{a query object or name of a query}

\item{...}{parameter values}

\item{query_location}{the location of the query}

\item{include_header}{if true, the header will be included in the sql query}

\item{append_params}{should the parameter values be appended to the
query in a comment?}
}
\value{
a formatted query
}
\description{
Replace parameters in query
}
