% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_helpers.R
\name{comma_join}
\alias{comma_join}
\title{Join a character vector together with commas}
\usage{
comma_join(s, leading_comma = FALSE, trailing_comma = FALSE, quote = FALSE)
}
\arguments{
\item{s}{a character vector to be joined}

\item{leading_comma}{should a leading comma be inserted?}

\item{trailing_comma}{should a trailing comma be inserted?}

\item{quote}{should the output be quoted?}
}
\value{
a length-1 character vector
}
\description{
Join a character vector together with commas
}
