% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{head.query_template}
\alias{head.query_template}
\title{Print the header of a query}
\usage{
\method{head}{query_template}(x, ...)
}
\arguments{
\item{x}{a query}

\item{...}{ignored}
}
\value{
silently returns the head
}
\description{
Print the header of a query
}
