% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_query}
\alias{load_query}
\title{Load a query from file}
\usage{
load_query(query_name, query_location = NULL)
}
\arguments{
\item{query_name}{name of the query. This can also be a path, in which
case query_location is ignored.}

\item{query_location}{location of the query}
}
\value{
loaded query
}
\description{
Load a query from file
}
