% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{query_from_string}
\alias{query_from_string}
\title{Build a query object from a length-1 character vector}
\usage{
query_from_string(s)
}
\arguments{
\item{s}{character containing query}
}
\value{
convert string into query object
}
\description{
Build a query object from a length-1 character vector
}
