% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querying.R
\name{query_substitute}
\alias{query_substitute}
\title{Replace parameters in query}
\usage{
query_substitute(
  q,
  ...,
  query_location = queries_location(),
  append_params = TRUE
)
}
\arguments{
\item{q}{a query object or name of a query}

\item{...}{parameter values}

\item{query_location}{the location of the query}

\item{append_params}{should the parameter values be appended to the
query in a comment?}
}
\value{
a formatted query
}
\description{
Replace parameters in query
}
